pip_requirements.txt
# Core pipeline runtime
requests
pyyaml
numpy
soundfile
torch
qwen_tts
apple-foundation-models

# local-llm-wrapper runtime notes:
# - Ollama transport uses stdlib only.
# - Apple transport now requires apple-foundation-models.
